Changing the type of id in the Guid, and adding a ProductInfo page

+++ ExampleOnlineShop.HttpApiClient.Test1/UnitTest1.cs
@@ -9,10 +9,11 @@ public class UnitTest1
     {
         
         var shopClient = new ShopClient("https://localhost:7003");
-        var product = new Product( 15, "mil", 100m, 100, "bla-bla", "/");
+        var product = new Product( Guid.NewGuid(), "mil", 100m, 100, "bla-bla", "/");
         //await shopClient.DeleteProduct(product.Id);
         await shopClient.AddProduct(product);
         var responseProduct = await shopClient.GetProduct(product.Id);
+        if (responseProduct == null) throw new ArgumentNullException(nameof(responseProduct));
         await shopClient.DeleteProduct(product.Id);
         Assert.Equal(product,responseProduct);
